chore(DOM): Hits.js -> Hits.tsx
